refactor tpm and vault code so agents do not import other agents
